Add EventUtilities class
