+++ app/src/main/java/com/example/partyapp/ui/EventScreen.kt
@@ -61,6 +61,7 @@ import com.example.partyapp.data.entity.Event
 import com.example.partyapp.data.entity.User
 import com.example.partyapp.data.relation.UserAddEventCrossRef
 import com.example.partyapp.services.EventFactory
+import com.example.partyapp.services.EventUtilities
 import com.example.partyapp.services.ImageChooserService
 import com.example.partyapp.services.NotificationScheduler
 import com.example.partyapp.ui.components.AddButton
@@ -84,7 +85,6 @@ private lateinit var loggedUser: User
 private lateinit var viewModel: EventViewModel
 private var isEditing: Boolean = false
 private var selectedDateNumber: Int = 0
-private const val BASE_DATE_INT = 1000_00_00
 
 @Composable
 fun EventScreen(
@@ -336,12 +336,12 @@ private fun EventDateDetail(modifier: Modifier = Modifier) {
         )
         if (isEditingMode()) {
             val calendar = Calendar.getInstance()
-            if (event.day < BASE_DATE_INT) {
+            if (event.day < EventUtilities().BASE_DATE_INT) {
                 val dayValue = calendar.get(Calendar.YEAR) * 10000 + calendar.get(Calendar.MONTH) * 100 + calendar.get(Calendar.DAY_OF_MONTH)
                 updateEvent(event.copy(day = dayValue))
                 selectedDateNumber = event.day
             }
-            var date: String by remember { mutableStateOf(dateToStr(getEventDateTime())) }
+            var date: String by remember { mutableStateOf(dateToStr(EventUtilities().getEventDateTime(event))) }
             PartyDatePickerComponent(
                 text = date,
                 onDatePicked = { year, month, day ->
@@ -352,7 +352,7 @@ private fun EventDateDetail(modifier: Modifier = Modifier) {
                 }
             )
         } else {
-            val calendar = getEventDateTime()
+            val calendar = EventUtilities().getEventDateTime(event)
             Text(text = dateToStr(calendar),color = Color.White)
         }
     }
@@ -390,10 +390,7 @@ private fun EventTimeDetail(modifier: Modifier = Modifier) {
                 }
             )
         } else {
-            Text(
-                text = event.starts + "-" + event.ends,
-                color = Color.White
-            )
+            Text(text = event.starts + "-" + event.ends, color = Color.White)
         }
     }
 }
@@ -451,9 +448,9 @@ private fun Actions(
         modifier = Modifier.fillMaxWidth(),
         horizontalArrangement = Arrangement.spacedBy(10.dp)
     ) {
-        if (!isCreatedByCurrentUser()) {
+        if (!EventUtilities().isEventCreatedBy(event, loggedUser)) {
             AddEventButton(eventViewModel)
-        } else if (isNewEvent()) {
+        } else if (EventUtilities().isNewEvent(event)) {
             SaveDiscardBtns(eventViewModel, onBackToPrevPage)
         } else {
             DeleteEventButton(eventViewModel, onBackToPrevPage)
@@ -536,7 +533,7 @@ private fun addParticipation(eventViewModel: EventViewModel) {
 
 private fun addNotification(context: Context) {
     val scheduler = NotificationScheduler()
-    val calendar = getEventDateTime()
+    val calendar = EventUtilities().getEventDateTime(event)
     Log.d("DELAYED_NOTIF", "Set notification: ${calendar.time}, ${event.name} ")
     scheduler.scheduleNotification(
         context = context,
@@ -546,18 +543,6 @@ private fun addNotification(context: Context) {
     )
 }
 
-private fun checkEventValid() {
-    if (event.name.isEmpty()) {
-        throw IllegalStateException("Event name cannot be empty")
-    }
-    if (event.day < BASE_DATE_INT && selectedDateNumber != 0) {
-        updateEvent(event.copy(day = selectedDateNumber))
-    }
-    if (getEventDateTime().before(Calendar.getInstance())) {
-        throw IllegalStateException("Event date cannot be in the past")
-    }
-}
-
 private fun saveNewEvent(
     context: Context,
     eventViewModel: EventViewModel,
@@ -565,7 +550,7 @@ private fun saveNewEvent(
     onBackToPrevPage: () -> Unit = {}
 ) {
     try {
-        checkEventValid()
+        EventUtilities().checkEventValid(event)
         val newID = events.map { it.eventId }
             .ifEmpty { listOf(0) }
             .max()
@@ -581,27 +566,7 @@ private fun saveNewEvent(
 }
 
 private fun isEditingMode(): Boolean {
-    return isNewEvent() || isEditing
-}
-
-private fun isNewEvent(): Boolean {
-    return event.eventId == -1
-}
-
-private fun isCreatedByCurrentUser(): Boolean {
-    return event.creator.username == loggedUser.username
-}
-
-private fun getEventDateTime(): Calendar {
-    val calendar = Calendar.getInstance()
-    val (h, m) = event.starts.split(":").map { it.toInt() }
-    calendar.set(Calendar.HOUR_OF_DAY, h)
-    calendar.set(Calendar.MINUTE, m)
-    calendar.set(Calendar.SECOND, 0)
-    calendar.set(Calendar.DAY_OF_MONTH, event.day.mod(100))
-    calendar.set(Calendar.MONTH, (event.day / 100).mod(100))
-    calendar.set(Calendar.YEAR, event.day / 10000)
-    return calendar
+    return EventUtilities().isNewEvent(event) || isEditing
 }
 
 private fun updateEvent(updated: Event) {
